refactor(ui): ♻️ split main.rs into focused modules
Before: one 1068-line file.
After: 9 modules with clear responsibilities.

  main.rs        — entry point, ErrorApp, eframe::App impl (toolbar/tabs)
  app.rs         — WatsonApp struct, constructor, refresh, mutations
  types.rs       — Tab, EditState
  colors.rs      — CLR_* constants
  format.rs      — fmt_*, parse_*, collect_projects, parse_tags
  widgets.rs     — shared UI helpers (project_autocomplete, date_filter_bar,
                   feedback_label, empty_frames)
  tab_log.rs     — show_log, show_edit_modal
  tab_add.rs     — show_add
  tab_report.rs  — show_report, render_epic_report, render_project_report

Duplicates removed:
- Project autocomplete popup (toolbar + Add tab) → widgets::project_autocomplete
- Date filter bar (Log + Report) → widgets::date_filter_bar
- Feedback label (toolbar + Add tab) → widgets::feedback_label
- Empty state placeholder (Log + Report) → widgets::empty_frames
