+++ rs_watson_ui/src/colors.rs
@@ -1,0 +1,7 @@
+use eframe::egui;
+
+pub(crate) const CLR_GREEN: egui::Color32 = egui::Color32::from_rgb(80, 200, 120);
+pub(crate) const CLR_RED: egui::Color32 = egui::Color32::from_rgb(220, 80, 80);
+pub(crate) const CLR_PURPLE: egui::Color32 = egui::Color32::from_rgb(180, 100, 220);
+pub(crate) const CLR_CYAN: egui::Color32 = egui::Color32::from_rgb(100, 180, 220);
+pub(crate) const CLR_YELLOW: egui::Color32 = egui::Color32::from_rgb(220, 180, 80);
